Improve parsing of lang=interp values
In the case of a misconfigured g:medieval_langs with, for example, a
missing interpreter value (`foo=`) the prior implementation would throw
a 'list index out of range error`. This implementation ensures that
never happens: it only calls `split()` if the medeival_langs entry has
an equals sign, and it always returns the "last element" from the split
array, which is always guaranteed to exist.