Use scratch buffer instead of preview window
The name of the tempfile in the statusline kind of bothered me and the
preview window didn't really feel like the best solution for this
anyway. Using a scratch buffer works a bit better and also allows
multiple command outputs to be displayed simultaneously (which the
preview window does not allow).