Correctly handle missing trailing newline (#12)
Nvim's jobstart callback appends an empty '' element to the buffered
output list indicating a trailing newline. However, if the output does
not have a trailing newline there is no empty element, so it is
incorrect to unconditionally remove the last element.

Instead, explicitly check that the final element is empty before
removing it.